Add cli-download-posts.js
This is rather rough, and doesn't really work nicely for non-technical users, but it does download posts from the given blog; the only challenge is dealing with API rate limits, currently it just crashes when it hits them, and then you have to update the `processTimestamp` line with the new parameters and re-run it in a while. But it works. It should hopefully also be compatible with the get-likes-media script.